refactors, tests; ref #MAT-3810
