+++ src/components/newMeasure/NewMeasure.tsx
@@ -8,15 +8,13 @@ import Measure from "../../models/Measure";
 
 import { getServiceConfig, ServiceConfig } from "../config/Config";
 import axios from "axios";
-import { useOktaAuth } from "@okta/okta-react";
-import { useOktaTokens } from "../../hooks/useOktaJwt";
+import useOktaTokens from "../../hooks/useOktaTokens";
 
 export default function NewMeasure() {
   const history = useHistory();
   const [measureList, setMeasureList] = useState<Measure[]>([]);
   const [serviceConfig, setServiceConfig] = useState<ServiceConfig>();
   const [serviceConfigErr, setServiceConfigErr] = useState<string>();
-  const { oktaAuth, authState } = useOktaAuth();
   const { getAccessToken } = useOktaTokens();
 
   if (!serviceConfig && !serviceConfigErr) {
@@ -47,20 +45,6 @@ export default function NewMeasure() {
         tw="mr-4"
         onClick={() => history.push("/measure/create")}
         data-testid="create-new-measure-button"
-      />
-      <Button
-        buttonTitle="Debug Okta"
-        tw="mr-4"
-        onClick={() => {
-          // console.log("okta: ", okta);
-          // eslint-disable-next-line no-console
-          console.log("okta auth: ", oktaAuth);
-          // eslint-disable-next-line no-console
-          console.log("auth state: ", authState);
-          // eslint-disable-next-line no-console
-          console.log("okta accessToken: ", getAccessToken());
-        }}
-        data-testid="debug"
       />
       <div tw="mx-5 my-8">
         <MeasureList measureList={measureList} />
